+++ lib/aruba/reporting.rb
@@ -50,7 +50,7 @@ if(ENV['ARUBA_REPORT_DIR'])
       end
 
       def output
-        @aruba_keep_ansi = true # We want the output coloured!
+        aruba.config.keep_ansi = true # We want the output coloured!
         escaped_stdout = CGI.escapeHTML(all_stdout)
         html = Bcat::ANSI.new(escaped_stdout).to_html
         Bcat::ANSI::STYLES.each do |name, style|