+++ History.md
@@ -355,6 +355,7 @@
 * Fix for Gemfile excludes windows for many gems (issue #282)
 * Make feature tests compatible with ruby 1.9.2
 * Gather disk usage for file(s) (issue #294)
+* Replace keep_ansi-config option by remove_ansi_escape_sequences-option
 
 ## [v1.0.0](https://github.com/cucumber/aruba/compare/v0.11.0...v1.0.0)